ModuleManagementServer/ai-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic-settings>=2.7